# frontend/openapi.json
{
  "components": {
    "schemas": {
      "CountrySummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetSummary"
            },
            "title": "Datasets",
            "type": "array"
          },
          "flag": {
            "title": "Flag",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "flag",
          "datasets"
        ],
        "title": "CountrySummary",
        "type": "object"
      },
      "DatasetIndexItem": {
        "properties": {
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "country_flag": {
            "title": "Country Flag",
            "type": "string"
          },
          "country_name": {
            "title": "Country Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "description",
          "source_name",
          "country_code",
          "country_name",
          "country_flag"
        ],
        "title": "DatasetIndexItem",
        "type": "object"
      },
      "DatasetSummary": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "description",
          "source_name"
        ],
        "title": "DatasetSummary",
        "type": "object"
      },
      "DatasetTablePayload": {
        "properties": {
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 del snapshot servido; null si se sirvió en vivo.",
            "title": "Fetched At"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Records",
            "type": "array"
          },
          "view": {
            "$ref": "#/components/schemas/DatasetViewSummary"
          }
        },
        "required": [
          "view",
          "records"
        ],
        "title": "DatasetTablePayload",
        "type": "object"
      },
      "DatasetViewSummary": {
        "properties": {
          "category_col": {
            "title": "Category Col",
            "type": "string"
          },
          "category_label": {
            "title": "Category Label",
            "type": "string"
          },
          "filter_cols": {
            "items": {
              "type": "string"
            },
            "title": "Filter Cols",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "category_col",
          "category_label"
        ],
        "title": "DatasetViewSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API JSON para dashboards de datos públicos abiertos. Versión estable: /api/v1.",
    "title": "Open Data Hub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/countries": {
      "get": {
        "deprecated": true,
        "operationId": "list_countries_api_countries_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CountrySummary"
                  },
                  "title": "Response List Countries Api Countries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Countries",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/crime/views": {
      "get": {
        "deprecated": true,
        "operationId": "list_crime_views_api_crime_views_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetViewSummary"
                  },
                  "title": "Response List Crime Views Api Crime Views Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Crime Views",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/crime/views/{view_key}": {
      "get": {
        "deprecated": true,
        "operationId": "get_crime_view_api_crime_views__view_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "view_key",
            "required": true,
            "schema": {
              "title": "View Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "nult",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 2,
              "title": "Nult",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTablePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Crime View",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/datasets": {
      "get": {
        "deprecated": true,
        "operationId": "list_datasets_api_datasets_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetIndexItem"
                  },
                  "title": "Response List Datasets Api Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/datasets/{country_code}/{dataset_key}/views": {
      "get": {
        "deprecated": true,
        "operationId": "list_dataset_views_api_datasets__country_code___dataset_key__views_get",
        "parameters": [
          {
            "in": "path",
            "name": "country_code",
            "required": true,
            "schema": {
              "title": "Country Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_key",
            "required": true,
            "schema": {
              "title": "Dataset Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetViewSummary"
                  },
                  "title": "Response List Dataset Views Api Datasets  Country Code   Dataset Key  Views Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dataset Views",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/datasets/{country_code}/{dataset_key}/views/{view_key}": {
      "get": {
        "deprecated": true,
        "operationId": "get_dataset_view_api_datasets__country_code___dataset_key__views__view_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "country_code",
            "required": true,
            "schema": {
              "title": "Country Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_key",
            "required": true,
            "schema": {
              "title": "Dataset Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "view_key",
            "required": true,
            "schema": {
              "title": "View Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "nult",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 2,
              "title": "Nult",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTablePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dataset View",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/health": {
      "get": {
        "deprecated": true,
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Api Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "deprecated"
        ]
      }
    },
    "/api/v1/countries": {
      "get": {
        "operationId": "list_countries_api_v1_countries_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CountrySummary"
                  },
                  "title": "Response List Countries Api V1 Countries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Countries",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_datasets_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetIndexItem"
                  },
                  "title": "Response List Datasets Api V1 Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/datasets/{country_code}/{dataset_key}/views": {
      "get": {
        "operationId": "list_dataset_views_api_v1_datasets__country_code___dataset_key__views_get",
        "parameters": [
          {
            "in": "path",
            "name": "country_code",
            "required": true,
            "schema": {
              "title": "Country Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_key",
            "required": true,
            "schema": {
              "title": "Dataset Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetViewSummary"
                  },
                  "title": "Response List Dataset Views Api V1 Datasets  Country Code   Dataset Key  Views Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dataset Views",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/datasets/{country_code}/{dataset_key}/views/{view_key}": {
      "get": {
        "operationId": "get_dataset_view_api_v1_datasets__country_code___dataset_key__views__view_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "country_code",
            "required": true,
            "schema": {
              "title": "Country Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_key",
            "required": true,
            "schema": {
              "title": "Dataset Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "view_key",
            "required": true,
            "schema": {
              "title": "View Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "nult",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 2,
              "title": "Nult",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "es",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTablePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dataset View",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Api V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "v1"
        ]
      }
    }
  }
}